Update Scorers (#5)

+++ grails/destressapp/grails-app/domain/org/ifcasl/destress/DurationScorer.groovy
@@ -4,7 +4,9 @@ class DurationScorer {
 
     //properties for each of the possible duration features store their weights
 
-    Float relativeSyllableDuration
-    Float relativeVowelDuration
+    Double relativeSyllableDuration
+    Double relativeVowelDuration
     //etc.
+
+    //constraints: weights must sum to 1 (see Scorer)
 }
