+++ grails/destressapp/grails-app/domain/org/ifcasl/destress/F0Scorer.groovy
@@ -4,7 +4,9 @@ class F0Scorer {
 
     //properties for each of the possible f0 features store their weights
 
-    Float relativeSyllableMean
-    Float relativeVowelMean
+    Double relativeSyllableMean
+    Double relativeVowelMean
     //etc.
+
+    //constraints: weights must sum to 1 (see Scorer)
 }
