+++ grails/destressapp/grails-app/domain/org/ifcasl/destress/IntensityScorer.groovy
@@ -4,7 +4,9 @@ class IntensityScorer {
 
     //properties for each of the possible intensity features store their weights
 
-    Float relativeSyllableMean
-    Float relativeVowelMean
+    Double relativeSyllableMean
+    Double relativeVowelMean
     //etc.
+
+    //constraints: weights must sum to 1 (see Scorer)
 }
